src/requirements.txt
# Install with: `pip install -r requirements.txt`

# Infrastructure
Django >= 1.6.5
Pillow==2.5.0
MySQL-python==1.2.5
python-memcached==1.53
South==1.0

# CMS libraries
django-fluent-blogs >= 0.9.7
django-fluent-pages >= 0.8.6
django-fluent-contents[code,text,oembeditem] >= 0.9
django-fluent-dashboard >= 0.3.6
django-fluent-comments >= 0.9.2
django-staff-toolbar >= 1.0

# Support libs
django-admin-tools >= 0.5.1, < 0.6
django-any-imagefield==0.8.1
django-any-urlfield>=1.0.10
django-axes==1.3.6
django-categories==1.2.2
django-crispy-forms==1.4.0
#dashboardmods >= 0.2.2
django_compressor==1.4
-e git+https://github.com/clintecker/django-google-analytics.git@140787ba86ef4995b2cabf927bf0b56ec7728348#egg=google_analytics-0.2
-e git+https://github.com/vdboor/django-filebrowser-no-grappelli-django14.git@fd921ac1014e3a923a703679c5576969fa191f5b#egg=django-filebrowser-no-grappelli-django14
django-taggit==0.12
django-taggit-autosuggest=0.2.4
django-threadedcomments==0.9.0
django-tinymce >= 1.5.2
django-wysiwyg == 0.7.0
html5lib >= 1.0b3
micawber >= 0.3.0
pytz>=2014.4
sorl-thumbnail==11.12.1b

# Debugging
django-debugtools>=1.1.2
bpython>=0.13
